refactor: move `core/types` to `plugin/evm/types`
