+++ eth/filters/filter.go
@@ -33,7 +33,7 @@ import (
 	"math/big"
 
 	"github.com/ava-labs/coreth/core/bloombits"
-	customtypes "github.com/ava-labs/coreth/core/types"
+	customtypes "github.com/ava-labs/coreth/plugin/evm/types"
 	"github.com/ava-labs/coreth/rpc"
 	"github.com/ava-labs/libevm/common"
 	"github.com/ava-labs/libevm/core/types"
