+++ eth/gasprice/fee_info_provider.go
@@ -31,7 +31,7 @@ import (
 	"math/big"
 
 	"github.com/ava-labs/coreth/core"
-	customtypes "github.com/ava-labs/coreth/core/types"
+	customtypes "github.com/ava-labs/coreth/plugin/evm/types"
 	"github.com/ava-labs/coreth/rpc"
 	"github.com/ava-labs/libevm/core/types"
 	lru "github.com/hashicorp/golang-lru"
