+++ ethclient/ethclient.go
@@ -35,7 +35,7 @@ import (
 	"math/big"
 
 	"github.com/ava-labs/coreth/accounts/abi/bind"
-	customtypes "github.com/ava-labs/coreth/core/types"
+	customtypes "github.com/ava-labs/coreth/plugin/evm/types"
 	"github.com/ava-labs/coreth/interfaces"
 	"github.com/ava-labs/coreth/rpc"
 	"github.com/ava-labs/libevm/common"
