+++ plugin/evm/types/hashing.go
